azure-key-vault/requirements.txt
azure-common           1.1.27
azure-core             1.18.0
azure-identity         1.6.1
azure-keyvault-secrets 4.3.0